requirements.txt
tensorflow
numpy
opencv-python
matplotlib
scikit-learn
pandas
streamlit
pillow
